Refactor parser person packages
