+++ src/test/java/peoplesoft/logic/commands/person/PersonEditCommandTest.java
@@ -1,4 +1,4 @@
-package peoplesoft.logic.commands;
+package peoplesoft.logic.commands.person;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
@@ -18,7 +18,7 @@ import org.junit.jupiter.api.Test;
 
 import peoplesoft.commons.core.Messages;
 import peoplesoft.commons.core.index.Index;
-import peoplesoft.logic.commands.person.PersonEditCommand;
+import peoplesoft.logic.commands.ClearCommand;
 import peoplesoft.logic.commands.person.PersonEditCommand.EditPersonDescriptor;
 import peoplesoft.model.AddressBook;
 import peoplesoft.model.Model;
