+++ src/test/java/peoplesoft/logic/commands/person/PersonFindCommandTest.java
@@ -1,4 +1,4 @@
-package peoplesoft.logic.commands;
+package peoplesoft.logic.commands.person;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -14,7 +14,6 @@ import java.util.Collections;
 
 import org.junit.jupiter.api.Test;
 
-import peoplesoft.logic.commands.person.PersonFindCommand;
 import peoplesoft.model.Model;
 import peoplesoft.model.ModelManager;
 import peoplesoft.model.UserPrefs;
