+++ src/test/java/peoplesoft/logic/commands/person/PersonListCommandTest.java
@@ -1,4 +1,4 @@
-package peoplesoft.logic.commands;
+package peoplesoft.logic.commands.person;
 
 import static peoplesoft.logic.commands.CommandTestUtil.assertCommandSuccess;
 import static peoplesoft.logic.commands.CommandTestUtil.showPersonAtIndex;
@@ -8,7 +8,6 @@ import static peoplesoft.testutil.TypicalPersons.getTypicalAddressBook;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import peoplesoft.logic.commands.person.PersonListCommand;
 import peoplesoft.model.Model;
 import peoplesoft.model.ModelManager;
 import peoplesoft.model.UserPrefs;
