+++ src/test/java/peoplesoft/logic/parser/job/JobFindCommandParserTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.api.Test;
 import peoplesoft.logic.commands.job.JobFindCommand;
 import peoplesoft.model.job.JobContainsKeywordsPredicate;
 
-public class JobPersonFindCommandParserTest {
+public class JobFindCommandParserTest {
 
     private JobFindCommandParser parser = new JobFindCommandParser();
